+++ packages/terra-clinical-result/tests/wdio/flowsheet-result-cell-spec.js
@@ -64,4 +64,10 @@ Terra.describeViewports('FlowsheetResultCell', ['medium'], () => {
 
     Terra.validates.element('entered in error');
   });
+
+  it('displays a FlowsheetResultCell in a semantic table with multiple padding styles', () => {
+    browser.url('/#/raw/tests/terra-clinical-result/clinical-result/flowsheet-result-cell/semantic-paddingstyle-with-flowsheet-result-cell');
+
+    Terra.validates.element('semantic table paddingStyles');
+  });
 });