+++ packages/terra-clinical-result/CHANGELOG.md
@@ -8,6 +8,9 @@
   * Changed ResultTimeHeaderCell to remove the `<time>` html tag for the date and time. This was causing issues and technically does not have full screen reader support.
   * Locked `uuid` dependency to `7.0.3` for consistency across Terra packages. 
 
+* Added
+  * Added site documentation for Flowsheet, Time Header, and Name Header cells usage.
+
 ## 1.18.0 - (June 22, 2023)
 
 * Added